+++ packages/ui/modules/index.d.ts
@@ -2,7 +2,7 @@
 
 import {ElementType, ReactElement, Factory} from "react";
 import {
-  Config, Conjunctions, Operators, Widgets, Types, Fields, Funcs, Settings as CoreSettings, CoreConfig,
+  Conjunctions, Types, Fields, Funcs, CoreConfig, RuleValue,
   InputAction,
   ImmutableTree,
   Actions,
@@ -11,15 +11,118 @@ import {
   ItemProperties,
   ValueSource,
   TypedValueSourceMap,
-  ConjsProps, FieldProps,
-  WidgetProps, TextWidgetProps, DateTimeWidgetProps, BooleanWidgetProps, NumberWidgetProps, SelectWidgetProps, 
-  TreeSelectWidgetProps, RangeSliderWidgetProps, CaseValueWidgetProps, ConfigContext, FactoryWithContext,
-  RenderedReactElement, SerializedFunction,
-  Utils as CoreUtils
+  ConfigContext, FactoryWithContext, RenderedReactElement, SerializedFunction,
+  ConjsProps,
+
+  ImmutableList, ImmutableMap, ImmutableOMap,
+
+  // to extend
+  Config as CoreConfigType,
+  Settings as CoreSettings,
+  Utils as CoreUtils,
+  // to override <C>
+  OperatorProximity as _OperatorProximity,
+  Operator as _Operator,
+  Operators as _Operators,
+  FieldProps as _FieldProps,
+  Widget as _Widget,
+  Widgets as _Widgets,
+  BaseWidget as _BaseWidget,
+  RangeableWidget as _RangeableWidget,
+  TypedWidget as _TypedWidget,
+  TextWidget as _TextWidget,
+  NumberWidget as _NumberWidget,
+  RangeSliderWidget as _RangeSliderWidget,
+  DateTimeWidget as _DateTimeWidget,
+  TreeSelectWidget as _TreeSelectWidget,
+  TreeMultiSelectWidget as _TreeMultiSelectWidget,
+  SelectWidget as _SelectWidget,
+  MultiSelectWidget as _MultiSelectWidget,
+  BooleanWidget as _BooleanWidget,
+  FieldWidget as _FieldWidget,
+  FuncWidget as _FuncWidget,
+  CaseValueWidget as _CaseValueWidget,
+  ProximityProps as _ProximityProps,
+  ProximityOptions as _ProximityOptions,
+  WidgetProps as _WidgetProps,
+  TextWidgetProps as _TextWidgetProps,
+  DateTimeWidgetProps as _DateTimeWidgetProps,
+  BooleanWidgetProps as _BooleanWidgetProps,
+  NumberWidgetProps as _NumberWidgetProps,
+  RangeSliderWidgetProps as _RangeSliderWidgetProps,
+  SelectWidgetProps as _SelectWidgetProps,
+  MultiSelectWidgetProps as _MultiSelectWidgetProps,
+  TreeSelectWidgetProps as _TreeSelectWidgetProps,
+  TreeMultiSelectWidgetProps as _TreeMultiSelectWidgetProps,
+  CaseValueWidgetProps as _CaseValueWidgetProps,
+  CoreOperators as _CoreOperators,
+  CoreWidgets as _CoreWidgets,
 } from "@react-awesome-query-builder/core";
 
 // re-export
 export * from "@react-awesome-query-builder/core";
+
+/////////////////
+// override <C> in types
+/////////////////
+
+export type OperatorProximity<C = Config> = _OperatorProximity<C>;
+export type Operator<C = Config> = _Operator<C>;
+export type Operators<C = Config> = _Operators<C>;
+export type FieldProps<C = Config> = _FieldProps<C>;
+export type Widget<C = Config> = _Widget<C>;
+export type Widgets<C = Config> = _Widgets<C>;
+export type BaseWidget<C = Config> = _BaseWidget<C>;
+export type RangeableWidget<C = Config> = _RangeableWidget<C>;
+export type TypedWidget<C = Config> = _TypedWidget<C>;
+export type TextWidget<C = Config> = _TextWidget<C>;
+export type NumberWidget<C = Config> = _NumberWidget<C>;
+export type RangeSliderWidget<C = Config> = _RangeSliderWidget<C>;
+export type DateTimeWidget<C = Config> = _DateTimeWidget<C>;
+export type TreeSelectWidget<C = Config> = _TreeSelectWidget<C>;
+export type TreeMultiSelectWidget<C = Config> = _TreeMultiSelectWidget<C>;
+export type SelectWidget<C = Config> = _SelectWidget<C>;
+export type MultiSelectWidget<C = Config> = _MultiSelectWidget<C>;
+export type BooleanWidget<C = Config> = _BooleanWidget<C>;
+export type FieldWidget<C = Config> = _FieldWidget<C>;
+export type FuncWidget<C = Config> = _FuncWidget<C>;
+export type CaseValueWidget<C = Config> = _CaseValueWidget<C>;
+export type ProximityProps<C = Config> = _ProximityProps<C>;
+export type ProximityOptions<C = Config> = _ProximityOptions<C>;
+export type WidgetProps<C = Config> = _WidgetProps<C>;
+export type TextWidgetProps<C = Config> = _TextWidgetProps<C>;
+export type DateTimeWidgetProps<C = Config> = _DateTimeWidgetProps<C>;
+export type BooleanWidgetProps<C = Config> = _BooleanWidgetProps<C>;
+export type NumberWidgetProps<C = Config> = _NumberWidgetProps<C>;
+export type RangeSliderWidgetProps<C = Config> = _RangeSliderWidgetProps<C>;
+export type SelectWidgetProps<C = Config> = _SelectWidgetProps<C>;
+export type MultiSelectWidgetProps<C = Config> = _MultiSelectWidgetProps<C>;
+export type TreeSelectWidgetProps<C = Config> = _TreeSelectWidgetProps<C>;
+export type TreeMultiSelectWidgetProps<C = Config> = _TreeMultiSelectWidgetProps<C>;
+export type CaseValueWidgetProps<C = Config> = _CaseValueWidgetProps<C>;
+export type CoreOperators<C = Config> = _CoreOperators<C>;
+export type CoreWidgets<C = Config> = _CoreWidgets<C>;
+
+/////////////////
+// extend config
+/////////////////
+
+export interface Config extends CoreConfigType {
+  conjunctions: Conjunctions,
+  operators: Operators,
+  widgets: Widgets,
+  types: Types,
+  fields: Fields,
+  funcs?: Funcs,
+  settings: Settings,
+}
+
+export interface BasicConfig extends CoreConfig {
+  settings: Settings,
+  operators: CoreOperators<Config>,
+  widgets: CoreWidgets<Config>,
+}
+
 
 ////////////////
 // common
@@ -30,6 +133,7 @@ type AnyObject = object;
 type TypedMap<T> = {
   [key: string]: T;
 }
+type Empty = null | undefined;
 
 
 ////////////////
@@ -127,6 +231,39 @@ export interface RuleErrorProps {
   error: string,
 }
 
+export interface RuleProps {
+  config: Config,
+  id: string, // id of rule
+  groupId: string | Empty, // id of parent group
+  parentField: string | Empty, //from RuleGroup
+  selectedField: string | Empty,
+  selectedOperator: string | Empty,
+  operatorOptions: AnyObject | Empty,
+  value: ImmutableList<RuleValue>, //depends on widget
+  valueError: ImmutableList<string>,
+  valueSrc: ImmutableList<ValueSource>,
+  asyncListValues: Array<any> | Empty,
+
+  isDraggingMe: boolean | Empty,
+  isDraggingTempo: boolean | Empty,
+  isLocked: boolean | Empty,
+  isTrueLocked: boolean | Empty,
+  reordableNodesCnt: number | Empty,
+  totalRulesCnt: number | Empty,
+  parentReordableNodesCnt: number | Empty,
+  onDragStart: Function | Empty,
+  handleDraggerMouseDown: Function | Empty,
+  removeSelf: Function | Empty,
+  confirmFn: Function | Empty,
+
+  //actions
+  setField(field: string): undefined;
+  setOperator(operator: string): undefined;
+  setOperatorOption(name: string, value: RuleValue): undefined;
+  setLock(lock: boolean): undefined;
+  setValue(delta: number, value: RuleValue, valueType: string, asyncListValues?: Array<any>): undefined;
+  setValueSrc(delta: number, valueSrc: ValueSource): undefined;
+}
 
 /////////////////
 // Settings
@@ -156,12 +293,12 @@ export interface RenderSettings {
   showLabels?: boolean,
   maxLabelsLength?: number,
   customFieldSelectProps?: AnyObject,
-  renderBeforeWidget?: FactoryWithContext<FieldProps> | SerializedFunction,
-  renderAfterWidget?: FactoryWithContext<FieldProps> | SerializedFunction,
-  renderBeforeActions?: FactoryWithContext<FieldProps> | SerializedFunction,
-  renderAfterActions?: FactoryWithContext<FieldProps> | SerializedFunction,
-  renderRuleError?: FactoryWithContext<RuleErrorProps> | SerializedFunction,
-  renderSwitchPrefix?: RenderedReactElement | SerializedFunction,
+  renderBeforeWidget?: Factory<RuleProps> | SerializedFunction,
+  renderAfterWidget?: Factory<RuleProps> | SerializedFunction,
+  renderBeforeActions?: Factory<RuleProps> | SerializedFunction,
+  renderAfterActions?: Factory<RuleProps> | SerializedFunction,
+  renderRuleError?: Factory<RuleErrorProps> | SerializedFunction,
+  renderSwitchPrefix?: Factory<AnyObject> | SerializedFunction,
   defaultSliderWidth?: string,
   defaultSelectWidth?: string,
   defaultSearchWidth?: string,
@@ -170,15 +307,6 @@ export interface RenderSettings {
 
 export interface Settings extends CoreSettings, RenderSettings {
 }
-
-/////////////////
-// Config
-/////////////////
-
-export interface BasicConfig extends CoreConfig {
-  settings: Settings,
-}
-
 
 /////////////////
 // ReadyWidgets
@@ -226,9 +354,11 @@ export interface Utils extends CoreUtils {
 
 export declare const Utils: Utils;
 
-/////////////////
+
+//////////////////
 
 export declare const Query: Query;
 export declare const Builder: Builder;
 export declare const BasicConfig: BasicConfig;
 export declare const VanillaWidgets: VanillaWidgets;
+